I need to stop these large single commits... CSV export was polished out. Menu mode was sort of working. Might get rid of settings in it. Added an analysis of your connections mode but I have a nonetype error. Restructured the entire project trying to clean it up... I think I made a bigger mess.
